Release through changesets instead of a manual version bump
Every PR now carries a changeset, merging one opens a chore: release PR, and merging that PR is the release. Changesets versions npm/smuggle-cli and scripts/sync-versions.ts copies that number onto Cargo.toml, Cargo.lock, and the four platform packages, so nothing picks a version by hand anymore.

Release.yml gains a workflow_call entry because GitHub does not fire workflows for tags pushed with GITHUB_TOKEN, which is why the old bump workflow's tag push could not reliably kick off a build. Publish.yml now calls it directly with the tag it just created, and the GitHub release notes come from the new CHANGELOG section.
